Atualização da view para exibir um painel de log salvando em arquivo

+++ src/br/com/thiagopagonha/psnapi/ImageCache.java
@@ -1,37 +1,49 @@
 package br.com.thiagopagonha.psnapi;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.net.URL;
 
+import android.content.res.Resources;
 import android.graphics.Bitmap;
 import android.graphics.Bitmap.CompressFormat;
 import android.graphics.BitmapFactory;
 import android.os.Environment;
+import android.util.Log;
+
+import static br.com.thiagopagonha.psnapi.CommonUtilities.TAG;
 
 /**
  * Image cache resolver
  */
 public class ImageCache {
 
-	public static int getImage(String avatar) throws IOException {
-		//First create a new URL object 
-		URL url = new URL(avatar);
-
-		String fileName = md5sum(avatar);
+	public static int getImage(String avatar) {
+		try {
 		
-		//Next create a file, the example below will save to the SDCARD using JPEG format
-		File file = new File(Environment.getExternalStorageDirectory() +  fileName + ".png");
- 
-		//Next create a Bitmap object and download the image to bitmap
-		Bitmap bitmap = BitmapFactory.decodeStream(url.openStream());
+			//First create a new URL object 
+			URL url = new URL(avatar);
+	
+			String fileName = md5sum(avatar);
+			
+			File file = new File(Environment.getExternalStorageDirectory(), fileName + ".png");
 
-		//Finally compress the bitmap, saving to the file previously created
-		bitmap.compress(CompressFormat.PNG, 100, new FileOutputStream(file));
+			if(!file.exists()) {
+				//Next create a Bitmap object and download the image to bitmap
+				Bitmap bitmap = BitmapFactory.decodeStream(url.openStream());
 		
-		return -1;
+				//Finally compress the bitmap, saving to the file previously created
+				bitmap.compress(CompressFormat.PNG, 100, new FileOutputStream(file));
+			}
+			
+			return  Resources.getSystem().getIdentifier(file.getAbsolutePath(), null, null);
+		} catch(IOException io) {
+			Log.e(TAG, "Erro ao manipular os arquivos de imagem" );
+			Log.e(TAG, io.getMessage() );
+			
+			return R.drawable.ic_launcher;	
+		}
 	}
 	
 	private static String md5sum(String md5) {
